LPS-128860 Move the upgrade log appender to portal-upgrade-impl module
