+++ modules/apps/portal/portal-upgrade-impl/build.gradle
@@ -3,6 +3,8 @@ dependencies {
 
 	compileOnly group: "biz.aQute.bnd", name: "biz.aQute.bnd.annotation", version: "4.2.0"
 	compileOnly group: "com.liferay", name: "org.apache.felix.gogo.runtime", version: "1.1.0.LIFERAY-PATCHED-2"
+	compileOnly group: "com.liferay", name: "org.apache.logging.log4j", version: "2.14.1.LIFERAY-PATCHED-1"
+	compileOnly group: "com.liferay", name: "org.apache.logging.log4j.core", version: "2.14.1.LIFERAY-PATCHED-1"
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
 	compileOnly group: "org.osgi", name: "org.osgi.annotation.versioning", version: "1.1.0"
